Dpkg::OpenPGP::Backend::GnuPG: Fallback to use «gpg dearmor» if present
When verifying a signature, we were using the internal dearmor() method,
as that makes it possible to not have gpg around when we are using gpgv.

The problem is that the internal dearmor() method does not handle
concatenated ASCII Armor blocks, and might then fail to find the
certificates. When using gpgv this is only a partial regression as
we were previously not verifying at all on minimal systems where gpg
was not available. But when it is available, now that has regressed.

In the future we might require no concatenated ASCII Armor blocks,
but for now let's mitigate this regression.

Fixes: commit a11d7340da2f679fc673d92767dfefcbb3eb1e69
Reported-by: Sven Joachim <svenjoac@gmx.de> (on IRC)