Revert "arch: Add support for loong64 CPU"
This reverts commit 321e479ba142e6b05d1a27b50e4bb7900903fbf0.

This arch which was supposedly already defined, with an agreed ABI,
GNU triplet and dpkg multiarch paths, seems to want to redefine all
these. There is no clear rationale for why this is happening at this
point in time, while patches for at least gcc have been submitted
with no coordination (which ended up being applied).

The handling of this is less than ideal. Until the whole situation
gets clarified, the safest way is to revert the support. It can be
reintroduced once things clear out.

Ref: #1028654