Dpkg::Vendor::Debian: Add time64 feature in the future area
This feature (disabled by default) enables 64-bit time on 32-bit
architectures where their ABI does not define it by default. This
feature automatically enables the lfs one as it is required by it.

Do not enable time64 for the new 32-bit architectures that already have
it enabled by default as part of their ABI, that is «arc» and «or1k».

Closes: #1030159
Based-on-patch-by: Helge Deller <deller@gmx.de>