dpkg: Remove path_quote_filename() usage
The original problem for which this function was introduced in
commit 520ad3054094d91201089bf97cf3e73266b6e879, is no longer relevant
since commit 1a8d12285d650fa4bb13406a515228030d4d9ca1 where the varbuf
function doing vprintf was fixed to not loop infinitely. Changes in
commit 520ad3054094d91201089bf97cf3e73266b6e879, further changed the
code to not even pass the filenames to fd_fd_copy() anymore, and instead
call the ohshit() functions from the call sites.

We leave the function around for now, if this ends up not being useful
for anything it will be eventually removed.

Warned-by: cppcheck
Changelog: internal