libdpkg: Move parse_nv() next word check inside else block
If we found a name/value match, and are using a fallback value, then we
will always have a NULL str_end, and the check for trailing junk will
never trigger. This check only makes sense in the else block where
we have checked for any non-space characters after the parsed text.