doc: Use «honor» and «behavior» uniformly in documentation and output
Use the American English variants uniformly. Leave the ones found in
changelog entries for other authors than myself to avoid putting words
in their mouths. We also leave the only remaining «our»-ending word
for «favour» as that is part of the maintscripts interface, and to avoid
confusion we leave other non-interface references alone for consistency.