Dpkg::Gettext: Relicense from BSD-2 to GPL-2+
This file was originally copied from the Debconf::Gettext module from
debconf with a BSD-2 license, originally by Joey Hess, which got
modified by Nicolas François on import, and subsequently reworked by
me, and with a trivial change by Raphaël Hertzog.

With agreement from authors with original or non-trivial code:

  Joey Hess
    <https://lists.debian.org/debian-dpkg/2023/03/msg00005.html>
  Nicolas François
    <https://lists.debian.org/debian-dpkg/2023/03/msg00002.html>

and myself with this commit, relicense this file from BSD-2 to GPL-2+
to simplify the licensing for the Dpkg perl modules.

Signed-off-by: Guillem Jover <guillem@debian.org>