dselect: Relicense ftp method from GPL-2 to GPL-2+
This code was merged from the dpkg-ftp package, which had a GPL-2
license.

With the agreement from authors with original or non-trivial code:

  Andy Guy
    <https://lists.debian.org/debian-dpkg/2023/03/msg00007.html>
  Christian Hudon
    <https://lists.debian.org/debian-dpkg/2023/03/msg00004.html>
  Yann Dirson
    <https://lists.debian.org/debian-dpkg/2023/03/msg00008.html>
  Martin Schulze
    <https://lists.debian.org/debian-dpkg/2010/04/msg00055.html>
  Raphaël Hertzog
    <https://lists.debian.org/debian-dpkg/2010/04/msg00054.html>

and myself with this commit, relicense this code from GPL-2 to GPL-2+
to simplify the licensing for the dselect methods and make it possible
to refactor and shuffle code around without creating licensing issues.

Signed-off-by: Guillem Jover <guillem@debian.org>