Dpkg::Substvars: Handle exponential expansion gracefully
We should switch our excessive recursive prevention check to cover
exponential expansion. Because with exponential expansion we always
make progress, the current check misses those cases, so we track
expansion per variable, and ignore those variables that do not contain
any dollar symbol as these cannot then be involved in further expansion.

In this context this is not considered a security issue, but a
robustness issue, where we do not want the code to end up consuming
boundless amounts of memory.