build: Add support for .git-blame-ignore-revs file
The current codebase contains a mix of styles, which is confusing when
reading, or when contributing code. The main reason the code has been
switched to the new style incrementally has been because mass changes
create «git blame» barriers, at which point git blame becomes unusable.

But with the recent-ish support for the «git blame» ignore-revs-file
we can now track these commits and let git ignore them when doing blame,
which removes the main reason for avoiding these mass changes.