libdpkg: Mask the mode from the tar entry to only the file mode bits
Depending on the tar implementation used, the mode in the tar entry
might include more than the file mode bits, which can cause unexpected
values to be set on the unpacked object. Mask them to make sure we are
dealing with known metadata.

This is currently a problem with AIX's tar, but could happen on other
systems using various tar implementations, or with .deb created with
builders other than dpkg-deb.