libdpkg: Simplify progress output by using fputs() and putchar()
Instead of doing a full printf(), when we know that we do not need to
actually format the strings, use the more efficient fputs() and
putchar() functions.

This also decouples the percentage printing which is locale sensitive
from the general format of the output, which we will then be able to
mark for translation independently of the other text.