dpkg-db-keeper: Add helper script to assist with tracking database changes
A simplified version of this script has been in constant use to assist
in dpkg development for years now, so that changes performed to the db
can be easily tracked and studied, when running dpkg versions from git.

But it is also generally helpful for record tracking of what has changed
over time in the system.