build: Remove vim swap file entry from .gitignore
This is a developer specific tool artifact that does not belong in the
project .gitignore file, as otherwise we'd be polluting these with
endless entries for any such tool that might be used locally. Instead
users should really be adding them to their ~/.config/git/ignore file
(or equivalent) for any such local tool generated artifact to be
ignored globally on those systems.