test: Pass -n to gzip for all invocations
Reduce potential for unreproducible output in the generated artifacts,
so that we do not need to take that into account in the tests, even
though this is currently not a cause for test failures. This also makes
these calls consistent with other gzip usage in the test suite which
were already using this option.