test: Compute the sizes for «dpkg-deb -I» output dynamically
The sizes of the parts composing the various .deb can change depending
on the specific compression library or command used, so hardcoding them
in the test means these are not robust and can fail when these change
their output, or when we test against different versions. Compute the
sizes dynamically during the test and use those as part of the expected
output.

For uncompressed members we do not need to parametrize these sizes.