Dpkg::Vendor::Debian: Invert time64 suppression list logic
New 32-bit glibc architectures default to time64, and the list of time32
architectures is finite and will not grow any longer. So it makes more
sense to base the ignore list on that finite list than on the new 32-bit
architectures that default to time64, even though currently that is a
shorter list, but which will require continue maintenance and update.