dpkg: Handle non-existent .dpkg-tmp files on read-only filesystems
On read-only filesystems trying to rename a directory or a file will
fail with EROFS, even if non-existent. To avoid impacting the common
case, check whether errno is EROFS, then check whether the file exists
and if so restore errno to EROFS to fail on that, otherwise let the
errno from the failure trickle into the next check which will ignore it
if it implies that the file does not exist.

Closes: #1030149