Dpkg::Vendor::Debian: Move lfs from future to abi area and create an alias
For the same reason we have moved the time64 feature from the future
to the abi feature area to make the name more clear and pass the test
of time once these start to be enabled by default on some architectures.
Do the same for the old lfs feature, but add a backwards compatibility
alias so that existing users of this interface do not break.

Suggested-by: Steve Langasek <vorlon@debian.org>