test: Fix t-conffile-divert-conffile case
We were renaming the conffile during the diversion addition, so when
configuring the package dpkg considered that a local admin change,
and prompted as expected. Switch to make a copy of the diverted file
and copy it back so that it does not disappear, and we will switch it
to the diversion copy method once that's available.