build: Do not enable the sanitizer for the functional test suite
The way the code is currently structured, it is not freeing memory
on code paths leading to an ohshit() or ohshite(). But those might
continue elsewhere if an error context has been configured to use
setjmp(). We will need to fix these instances before this can be
enabled again on the CI.

This got botched test locally due to overridden CFLAGS usage.

Fixes: commit 59268c3f1098114c0dc27b2129882faa96897a1a
Changelog: silent