dpkg-buildtree: New program
This is a new helper program that initially will try to abstract
operations that deal with internal dpkg implementation details where
packagers in general should not need to be concerned about.

The first operation introduced is to clean state files generated by
various dpkg tools, so that we can eventually move their locations
through a new dpkg-build-api level for example.