debian: Set Persistent=true for dpkg-db-backup systemd timer
The "OnCalendar=daily" rule means every night exactly at 00:00:00. The
timer does not trigger if the computer is not turned on exactly at that
time.

The timer is meant to trigger daily even if the computer is not running
at midnight. By setting Persistent=true we make the timer trigger
immediately if the specified time has passed.

Closes: #1041730
Signed-off-by: Guillem Jover <guillem@debian.org>