libdpkg: Add support for sysuser/sysgroup specific to each chroot
When operating on a chroot via --root, we should always parse the
passwd and group files from within that chroot, and not from the
host system.

We now parse the passwd and group files explicitly, so that we can
use the pathname of our choosing. The defaults are /etc/passwd and
/etc/group which can be overridden with the new environment variables
DPKG_PATH_PASSWD and DPKG_PATH_GROUP, all relative to the --root
pathname. Although we currently hardcode these paths, which we
previously delegated to the libc, these are pretty standard locations
and this was not working for chroots anyway. We might eventually make
this selectable at configure time, but for now the defaults and the
environment variables should be good enough.

Closes: #1033051