man: Use semantic line breaks (must requirements)
Add line breaks after every «.», «!» and «?» that ends a sentence. We
only reflow or fold lines that are getting modified anyway as part of
this change, to avoid diff churn.

In addition this will make the manual page output consistent, and removes
duplicate paragraphs in the translations only differing the spacing after
the full stop.

This change should not affect the semantics of the sentences and should
produce a no-change «git diff --word-diff» output.

See <https://sembr.org/>.