test: Add installtest target and make test use the build tree programs
The usual convention is for a test or check target to use the programs
from the build tree, and an installtest or installcheck target to test
the as installed programs from the system.

Set the DPKG_BUILDTREE variable automatically for test and add a new
installtest target to make the usage more standard and easier to use.