dpkg-split: Revert switch to use DEBEXT for --help output
We cannot use string concatenation via a macro in strings marked for
translation that end up being parsed by xgettext, as the resulting
msgid strings end up truncated.

Restore the previous hardcoding, as using a format string makes the
string to translate harder to understand, and instead we'll probably
need to add a new macro to only contain the actual extension w/o the
dot and use it via a format string, and unify all other messages too.

Reported-by: Sven Joachim <svenjoac@gmx.de>
Fixes: commit db1de20385ff04579ad3cb421e3dcc62f1df13f4
Changelog: silent