build: Split sanitizer unit tests from author checks
For the same reason we recently reverted the sanitizer tests in
commit 90cadbc8698fc7e3ca0cf018a2534c8874ef7376 for the functional
side of the test suite, we should not be running any other tests
that are functional in nature.

Split the author checks to be the main full test suite driver, and
restrict the unit-tests to enable the sanitizer only for the lib/
unit tests.