Dpkg::Vendor::Debian: Fix stackclash condition to enable the feature
We were using the cpu to check where to enable this feature, but this
should have been the arch, as that's what the names we match against
are. We used these specific names to avoid enabling this feature on
all arm based architectures where this feature might not be supported,
and instead only enable it on new ones.

Closes: #1054583