test: Pass argument S to ar to avoid generating a symbol table
Depending on the implementation, the command might decide to add a
symbol table based on heuristics on the data we are adding. For example
with llvm-ar when adding a zeros-only file, it decides to add such
symbol table, which breaks the .deb format.

In the future we need to switch to a custom ar implementation when
generating test artifacts to avoid depending on system behavior or
diverging ar formats that are not compatible with the ones a .deb is
expected to use, such as the AIX Big and Small ar formats.