build: Add support for structured procfs via struct psinfo
On AIX and recent versions of Solaris and derivatives such as Illumos,
there is an API to access the process information via structured
data from /proc files.

On at least Solaris <= 11.3 this interface is not compatible with LFS, so
we need to check whether the header is available and usable, and then for
the code using structured procfs we need to disable LFS when needed.