build: Only link against libtinfo if needed
If the check determines the symbol we look for is already available,
then we should not link against the libtinfo library unconditionally.
Instead move the assignment outside the AC_SEARCH_LIBS, and check
for whether the function is not available at all or is builtin, and
only if these do not apply link against the library autoconf has
determined to use.

Fixes: commit c03aeaa372a1507ec6835c8613388f84c56228d1
Changelog: silent