test: Improve cppcheck coverage
We needed to at least define __GNUC__ and __GNUC_MINOR__ macros,
otherwise cppcheck could not see the code attributes. And to point
to the include directories used in the code base to make the headers
analyzable.

We need to add a new bunch of suppressions, which we add inline for
false positives, that we do not want to track per line or where we do
not want to suppress on the entire file.

This makes it possible finally, to remove a bunch of suppression for
previous false positives due to lack of information for cppcheck.