u-a: Wrap the bad_format() function pointer call into a new function
This simplifies somewhat the callers which do not need to use ctx to
access the parse error function and pass it as an argument. And makes
it possible for cppcheck to understand that this is a non-returning
function, which it seems to be unable to do with a function pointer.