dpkg-shlibdeps: Do not require debian/control when parsing CLI options
Now that we are not registering the private library directories while
parsing the options, but instead we queue them for later. We can move
the parsing of debian/control, after the option parsing too, so that
we do not require it, which made the file to be required from the
current working directory even for options such as --version or --help.

Fixes: commit 292655415c34db23e6d5014b6f71d0060f5811e4
Closes: #1055536