arch: Restrict kopensolaris ports to amd64 and i386
This port got started but it does not look like it went very far,
but there is at least one semi-active project (OS Dyson) which
supposedly still has plans to continue this port, although there has
been no reply yet as to what are their plans or status. For now just
restrict the port to the CPUs that have support in the glibc fork.
We can always revisit whether to remove (with the option to add back)
the port in the future.

Ref: https://lists.debian.org/debian-dpkg/2023/11/msg00009.html
Ref: https://github.com/ip1981/kopensolaris-glibc