Add support for native arch definition for chroots
When operating over a non-default root directory, we need to switch the
native architecture and use the one from the dpkg installed in that
chroot, instead of our current one.

This information is available in the status database, but that would
require parsing it and hardcoding the package name dpkg is shipped,
which might not match the one used inside that chroot.

We could also ship a file with the native architecture, that then we can
load. But that implies hardcoding the pathname for a filesystem file
(in contrast to the database pathname which we need anyway and can
specify via --admindir), which is a distribution specific policy, and
might end up not being loaded if the appropriate pathname is not passed,
making the native architecture become out of sync with the one in the
database.

The approach selected as the better one is to generate a file in the
database directory during package installation, so that we also cover
cross-grades. In the future we might provide a new action or helper to
populate a chroot directory for foreign installations, so that this does
not need to be done by bootstrapping programs.

Closes: #825385, #1020533