Dpkg::Vendor::Debian: Mask PIE for alpha and ia64
These architectures have poor PIE support in the toolchain and are
causing hard to diagnose problems in packages, where the only option
is to disable PIE. Instead of having to conditionally do that per
package, mask it globally. If this ever changes we can revisit this,
or PIE can be enabled by default in the toolchain.

Closes: #1040062