Dpkg::Control::Types: Add aliases for type enumerations
Refer to the debian/control file as the template file (CTRL_TMPL_
prefix). Refer to .dsc and .deb as CTRL_DSC and CTRL_DEB, as using
the CTRL_PKG_ prefix is very confusing given that we use _PKG to
denote binary package types. Refer to the Packages and Sources
repository indices using the CTRL_REPO_ prefix, as the CTRL_INFO_
prefix is rather confusing.