debian: Install s-s-d under /usr/sbin
On systems that have a merged-/usr this makes sure to install the
program in its canonical location on disk. On derivatives that do not
use a merged-/usr layout we create backward compatibility symlinks,
which can be removed once there are no more references to the absolute
pathname or the one under /sbin.

This makes it possible to remove the special casing for s-s-d and its
destination on install.

Closes: #1059982