debian: Check for usrmerge in bug-script on the same vendors as postinst
Make both scripts emit the warning on the same vendors. We duplicate the
logic for now, because the code in postinst emits colors, and introducing
a new shell library is potentially problematic for the maintscript, as
that can end up as a forever interface. And this check is supposed to
disappear eventually. If the get_vendor function proves useful it can
always be made into a shell library.