debian: Do not install dpkg-fsys-usrunmess on Debian and Ubuntu
These distributions are now fully moving all pathnames in the shipped
.deb packages under /usr and do not support any other filesystem layout.

We keep shipping it for any derivative that has not opted-out.

Ideally we'd do this by leaving the entries in the debhelper fragment
files and then excluding them with -X when necessary, but that is not
supported by dh_installman, so we do the next available thing, which
is not great, but works.