libdpkg: Handle tar long GNU names and links not being NUL terminated
The GNU tar implementation includes a terminating NUL byte as part of
the body of the long GNU name and link entries, which gets represented
in the size in the tar header. But not all tar implementations include
the NUL byte as part of the body or it size.

Even though dpkg-deb will never generate such packages, we have to be
robust against such input, so we need to allocate one more byte for
the terminating NUL byte, and explicitly terminate the string to make
sure what we return is a valid NUL-terminated string, otherwise we end
up with read overruns.

Closes: #1061404
Stable-Candidate: 1.21.x 1.20.x