dpkg-buildpackage: Fix the debian/rules executable check to respect -R
When the user has specified a different rules file via the -R option,
we should try to use that file for the check that makes sure it is
executable. But, unfortunately, because the -R option might be passed
other program and arguments to interpret the file or something else
entirely, we can only be sure this is a single file to be executed if
it contains a single element.

This means the check might get disabled when passing more than one
argument, but because it's a non-fatal warning, it should be fine.