dpkg-buildpackage: Remove fallback handling for missing required targets
The fallback code that was auto-detecting for potentially missing
build-arch and build-indep targets has been there since 2012, and has
increasingly been restricted. In Debian there was a mass bug filing
and there's now only a handful of packages missing these. It has come
the time to remove this fallback code and fully require these targets,
which means we can now avoid the auto-detection and fallback code.