dpkg-query: Fix exit codes for --show
The manual page states that (most) dpkg-query commands will return 0 for
success, 1 for files or packages not being found, and 2 for fatal errors.
But the code was coercing the return values from each function into a
boolean, which meant that the --show command would not return 2 on parse
errors for the --showformat.

Move the coercion into each command function, and map the format error
in --show into a proper 2.

Closes: #1064036