Dpkg::Substvars: Add support for implicit substvars assigned with $=
This operator adds support for implicit substvars, which will
instantiate the field referred in their name if that field does not
exist.

These variables need to be namespaced with a name followed by a colon
and then the field they will instantiate. All namespaces get appended
using the known separator for their field. If the field is not known
then space is used as the separator.

This will make it possible to generate fields automatically by external
tools.