dpkg-shlibdeps: Add support for new --package option
This option takes care of changing some of the defaults based on the
information from the debian/control file.

We select whether to use the Pre-Depends or Depends field as the default
on whether the package is Essential:yes. We select the package type from
the Package-Type field. And add the package to the set of packages
excluded by -x.