test: Pass -T+1 to xz to workaround spurious warning with xz 5.6.0
This new xz version will automatically emit a warning when it reduces
the amount of threads to use if it exceeds the soft memory limit, even
when the user has not specified a number of threads to use. This is
making the test fail. Add an explicit -T+1 for now to force parallel
mode with a single thread until the warning problem gets sorted out in
xz upstream.

Closes: #1064856