Dpkg::Vendor::Debian: Unconditionally set qa bug-implicit-func
For the time64 default change, we conditionally enabled
bug-implicit-func to avoid silent ABI breakage due to implicit function
declarations that end up using the time32 symbols, but that is causing
confusion as the effects are not visible on the most commonly used
architectures. Instead, enable this globally, unless the maintainer has
specified otherwise.

Ref: #1065371