build: Rework subst handling for built or installed artifacts
Switch from sed to perl so that we can perform in-place substitutions in
a portable way. Make the variable naming usage uniform.

This reduces the amount of code duplication to deal with these
substitutions.

We do not use AM_V_GEN any longer for the subst_.*_file macros, as
those get the filename as an argument, and are used in install-data-hook
targets, which then generate very confusing output such as
«  GEN    install-data-hook» for every modified file. And for the
subst_.*_filter macros we use AM_V_GEN in the call sites, if needed,
as that might not be appropriate if called after a pipe.