dpkg: Use a variable for each conffile pathname type
Instead of reusing two variables for various purposes, which makes the
code hard to follow and is rather confusing, use a dedicated variable
for each pathname type to make this all obvious.

This also removes the rather unsafe code pattern of tracking a pointer
state to then append to with strcpy().

Changelog: internal