Revert "test: Pass -T+1 to xz to workaround spurious warning with xz 5.6.0"
This reverts commit 5822b37460127d772f06a09b0acceaf87a1fec66.

This got fixed in xz 5.6.1, but those two versions were backdoored
anyway. New version will not emit the warnings that made the test suite
fail, and we should not keep workarounds for versions that were
backdoored anyway.

The workaround was causing test failures on systems with older xz
version where the -T+1 option form is unknown.