debian: Suppress s-s-d compat symlink if /sbin is missing
When bootstrapping, if base-files has not yet been unpacked, the /sbin
pathname is going to be missing and we will not be able to create a
compat symlink, and «ln» will fail aborting the package installation.

We cannot unconditionally create the /sbin directory as that would break
on usr-merged systems that ship a symlink in a package such as base-files.

Ideally base-files would be unpacked first by the bootstrapping tool,
but that would required encoding this ordering somehow.

Closes: #1071078
Signed-off-by: Guillem Jover <guillem@debian.org>