build: Workaround Tap::Harness verbose misbehavior on parallel mode
The Tap::Harness module misbehaves when enabling both parallel and
verbose mode, where the verbose output is eaten by some concurrency
statistics output, and only the last test output is emitted.

We set the formatter_class explicitly to the file formatter in case we
are running in verbose and parallel modes as a workaround.

Ref: https://github.com/Perl-Toolchain-Gang/Test-Harness/issues/105