build: Add GitHub CI support for tests on ports
Add jobs for FreeBSD, NetBSD, OpenBSD, OmniOS (illumos) and macOS.

On NetBSD we install gsed, as the BSD sed does not like the
utf8toman.sed script.

We cannot enable Solaris, as the packages are very outdated, and
autopoint cannot find the archive.dir.tar.xz and fails to even
autoreconf. But the community based OmniOS based on illumos is a better
and more up-to-date replacement.

This uses the actions from <https://github.com/vmactions>.