Dpkg::Shlibs::Cppfilt: Normalize demangled symbols with llvm or C++11 format
The output from c++filt from llvm produces no spaces between ending
angle brackets (<<symbol>>) as allowed by C++11, contrary to what
GNU binutils does by default as it was not allowed before C++11.

Because this is used to compare against demangled symbols in symbols
files, we need to use an unified format, given that the space around
angle brackets is more compatible, and has been the default when using
GNU binutils, we normalize into that.