test: Add new DPKG_CHECK_DIFF macro to abstract file comparisons
On some systems, such as ones based on SunOS/illumos, the diff tool does
not honor the -q option and outputs «No differences encountered» when
there are no file differences.

Instead of requiring GNU diff, abstract the check into a new macro that
uses the internal autotest machinery to compare the files for us via the
expout mechanism.