dpkg-ar: New internal ar implementation script
This script is going to be used in the test suite to guarantee we
have an implementation that conforms to the common Unix ar format,
independently of what the system supports.

This is a minimal interface, which is intended to be used internally
for now, and might get replaced with a C implementation later on.

Changelog: internal