build: Add missing files and sort POTFILES.in
Let's add all files even if they currently do not contain any
translation, as not listing them all unconditionally is error prone,
as we might end up adding translations and those getting missed due
to the file already existing.

Also this way we can add an author check for this.