dpkg: Check for < 0 instead of == -1 for conffderef() return values
This is a dpkg specific function that returns 0 on success and -1 on
error. While a static analyzer should have enough in-tree information
to be able to avoid emitting bogus diagnostics, we switch to the same
pattern as we just did for syscall return values so that the coding
style is uniform.

While at it we improve the debug output for the newly handled values.

Changelog: internal