libdpkg: Fix execname support on GNU/Hurd
Implement the missing code, and handle fcntl() returning -1 for the
locked pid.

We need to link against Hurd's libps and add it to libdpkg.pc, for the
process functions.

We need to protect the <dpkg/file.h> inclusion when including <hurd.h>
as they have conflicting symbols for file_lock(), where the dpkg header
is only needed on Solaris-based systems.

Fixes: commit d6a9bb6951166a7254ab2b4f887c69c4bd4fb1a7