libdpkg: Add new varbuf_array() and C++ operator[] methods
We want to be able to replace current raw pointer usage, which accesses
those pointers as arrays. We add these methods so that we can switch the
code to use varbuf, while at the same time, making it safer to use, as
the new code will have built-in range checks.