Revert "scripts/mk: Reduce the number of subprocesses"
This partially reverts commit e146a68a375453db89bc3b5e816d9405e5fd5634.

While the reverted commit significantly reduced the amount of
subprocesses, it has caused some breakage on functionality packages were
relying on, some of which is currently hard to diagnose due to make
programming model complexity (at least the regressions for the build
flags), and another set is due to external packages using internal dpkg
macros. In any case the breakage is too big, and for some of the changes
potentially silent.

Keep the documentation changes from the original patch, and preserve the
newly introduced DEB_TIMESTAMP variable.

Revert this for now until the situation is more clear. And then it can
be reevaluated how to potentially reintroduce this, perhaps through a
new dpkg-build-api level, if that would require changes in packaging.

Closes: #1076904, #1077005