dpkg-buildtree: Add new command is-rootless
This new command can be used to query whether the build tree requires
root to build or can be built with no root privileges.

This is useful for build frontends that might need to know whether they
have to install additional gain-root-command packages to be able to
build the sources.

Suggested-by: Niels Thykier <niels@thykier.net>